==> source/cases/_R/insurance/rice/1_introduction.rmd <==
# Introduction

## Background

This case study discusses index based crop insurance in Northeastern Tanzania for rice paddy using time series satellite derived information. Use of satellite based metrics to design index insurance is gaining traction due to lower cost, potential for rapid implementation, historical data availability and unbiased estimate. However robust insurance contract that is both attractive to farmers and profitable for insurers, requires strong correlation between remote sensing measurements and yield loss. This is an active area of research within the remote sensing community and developing a 'good' yield prediction model is a challenging task, particularly for the smallholder systems.       
 
The study area is located in south-east part of the Kilimanjaro district in Tanzania, composed of the towns and villages of Maore, Kihurio, Ndungu, and Bendera (Ma-Ki-Ndu-Be). Rice is grown  grown once annually in independently managed small plots that are often clustered together. The main growing season is between mid-November and mid-March, which covers the short rains period between December and February. Rice yield varies with irrigation with highest yield in the Ndungu irrigation scheme zones. Northern zones have the highest drought risk as they depend on the rive water for irrigation. Fertilizier application is limited. More details about this project can be found in [Flatnes et al. 2019](https://basis.ucdavis.edu/sites/g/files/dgvnsk466/files/2018-11/Flatnes%20Carter%20conditional%20audit%20Tanzania.pdf).

In this exercise you will learn the workflow to design a contract based on satellite data. For the remote sensing analysis our goal is to  produce one of more seasonal values (metrics) for each plot location for which yield data are available. The seasonal variable can then be correlated with the yield variable to define the mapping function coefficients for yield index generation. Remote senisng analysis is followed by quality evaluation of contract desgined. Major steps are outlined below.

## Steps:

-  Download and pre-process MODIS data
-  Map rice and extract phenology
-  Create indices to predict yield
-  Model yield as a function of the indices
-  Design/evaluate contract


## Requirements 

We will primarily use three R packages for this exercise. `terra` and `luna` focus on raster processing and remote sensing data download respectively. `agrins` is a helper package to install all other packages and accessing project specific data. Here are [installation instructions](../installation.html).
