==> source/cases/_R/insurance/rice/6_area_yield.rmd <==

# Area yield insurance

## Introduction 

Here we consider an "area-yield" insurance contract that triggers if the average rice yield of an insurance zone is 10% below the zone's long-term rice yield. 

If yields are below this threshold, the contract makes payouts equivalent to the value of 90% of the zone's long-term average yield minus the observed zonal rice yield for that season. The actual amount paid to a household depends on the area insured by a household. 


## Basics

First a refresher. Below *y* is a zone's crop yield in season (time)  *t*; and *y_z* is the zone's long-term average rice yield. Yields are expressed in kg/ha. 

```{r ins10}
get_payout <- function(y, y_z, trigger, price){
   pmax( (trigger * y_z) - y,  0) * price
}
```

To see how that works. Assume we observed the following sequence of rice yields.

```{r ins11}
yields <- seq(500, 2500, 100)
mean_yield <- mean(yields)
yields
mean_yield
```

We set the rice price to $0.23 per kg.
```{r ins12}
price <- 0.23
```

Now we can compute payouts

```{r ins13}
payout <- get_payout(yields, mean_yield, trigger=0.9, price=price)
plot(yields, payout, type="l", xlab="yield", ylab="payout", las=1)
```

If insurance did not cost any money, income could be computed like this.

```{r ins112}
rice_income <- yields * price
total_income <- rice_income + payout
plot(yields, total_income, type="l", xlab="yield", ylab="income", las=1)
```

This shows that, thanks to insurance, income has a floor of $`r round(min(total_income))`. 

The certainty equivalent without this insurance (not considering the premium, or other sources of income) would be

```{r ins121, message=FALSE}
library(agrodata)
agro::ce_income(rice_income, rho=1.5)
```

And with insurance it would be
```{r ins212, message=FALSE}
agro::ce_income(total_income, rho=1.5)
```

A more realistic situation might be a household that has additional income of $100 per year and that the insurance costs $50 per year.

```{r ins20}
rice_income <- yields * price
insurance_income <- payout - 50
total_income <- rice_income + insurance_income
agro::ce_income(rice_income, rho=1.5)
agro::ce_income(total_income, rho=1.5)
```

The certainly-equivalent income with insurance is still higher. We can plot that:

```{r ins30}
plot(yields, rice_income, type="l", xlab="yield", ylab="income", las=1, ylim=c(-60,600), lwd=2)
lines(yields, insurance_income, col="blue", lty=2, lwd=2)
lines(yields, total_income, col="red", lwd=2)
legend("topleft", c("rice", "insurance", "total"), col=c("black", "blue", "red"), lty=c(1,2,1), lwd=2)
```

CE depends on rho

```{r ins40}
rho <- seq(0,3,.1)
ce_rice <- sapply(rho, function(r) agro::ce_income(rice_income, rho=r))
ce_tot <- sapply(rho, function(r) agro::ce_income(total_income, rho=r))
plot(rho, ce_rice, col="red", type="l", las=1, ylab="CE", lwd=2)
lines(rho, ce_tot, col="blue", lwd=2)
legend("topright", c("no inusrance", "insurance"), col=c("red", "blue"), lty=1, lwd=2 )

# certainty equivalents are the same at
i <- which.min(abs(ce_tot - ce_rice))
rho[i]
abline(v=rho[i], lty=3, col="gray")
```


## Tanzania households 

Now get the household data from the previous chapter.

```{r ins2}
z <- readRDS("hh_rice_yield.rds")
head(z)
```

Compute payouts per ha for each zone.

```{r ins3}
z$payout <- get_payout(z$y_zt, z$y_z, 0.9, price)
head(z)
```

Now we can compute an insurance premium (per ha) for each zone *z*. First we determine the *actuarially fair* price where the premiums paid are equal to the expected value of compensation (payouts) paid. 

```{r ins4}
pay_ha_year <- tapply(z$payout, z$year, mean)
pay_ha_year
afp <- mean(pay_ha_year)
afp
```

Let's assume a markup of 20% (per ha)

```{r ins444}
premium <- round(afp, 1) * 1.2
premium
```

## zonal premium

You could also compute premiums by zone 

```{r ins440}
zz <- aggregate(z[, "payout", drop=FALSE], z[, "zone", drop=FALSE], mean)
zz$premium <- zz$payout * 1.2
zz
```


## income 

For each household we can now compute income with and without insurance. 

```{r ins50, out.width = '100%'}
z$income <- z$y * price
z$income_with_ins <- z$income + z$payout - premium
head(z)
```


## certainty equivalents by household

```{r ins06, out.width = '100%'}
hh <- aggregate(z[, c("income", "income_with_ins")], z[,"fid", drop=FALSE], function(i) agro::ce_income(i, 1.5))

hh$benefit <- (hh$income_with_ins - hh$income)
hh$rel_benefit <- hh$benefit/ hh$income
head(hh)

mean(hh$benefit)
quantile(hh$benefit)
```

So, at the individual level, the contract meets the welfare test if you use the mean, but not if you use the median. 



## certainty equivalents by zone

```{r ins066, out.width = '100%'}
zz <- aggregate(z[, c("income", "income_with_ins")], z[,"zone", drop=FALSE], function(i) agro::ce_income(i, 1.5))

zz$benefit <- (zz$income_with_ins - zz$income)
zz$rel_benefit <- zz$benefit/ zz$income
head(zz)

mean(zz$benefit)
quantile(zz$benefit)
barplot(sort(zz$rel_benefit), ylab="relative benefit", xlab="zone", las=1)
```

So, at the zone level, the contract meets the welfare test both for the mean and the median (at rho=1.5).
